requirements.txt
dnspython==2.1.0
future==0.18.2
tqdm==4.61.0
